+++ web/js/datastore/DocMetaRef.ts
@@ -1,7 +1,9 @@
 /**
  * Represents a light weight reference to a DocMeta file.
  */
-import {DocInfo} from '../metadata/DocInfo';
+import {IDocInfo} from '../metadata/DocInfo';
+import {DocMeta} from '../metadata/DocMeta';
+import {Preconditions} from '../Preconditions';
 
 export interface DocMetaRef {
     readonly fingerprint: string;
@@ -15,6 +17,28 @@ export interface DocMetaFileRef extends DocMetaRef {
 
     readonly filename: string;
 
-    readonly docInfo: DocInfo;
+    readonly docInfo: IDocInfo;
+
+}
+
+export class DocMetaFileRefs {
+
+    public static createFromDocMeta(docMeta: DocMeta): DocMetaFileRef {
+
+        return this.createFromDocInfo(docMeta.docInfo);
+
+    }
+
+    public static createFromDocInfo(docInfo: IDocInfo): DocMetaFileRef {
+
+        Preconditions.assertPresent(docInfo.filename);
+
+        return {
+            fingerprint: docInfo.fingerprint,
+            filename: docInfo.filename!,
+            docInfo
+        };
+
+    }
 
 }